Refactor FilterPicker
