--- src/UI/Pages/CompileReports/FilterWidget/FilterPicker.tsx
@@ -1,48 +1,0 @@
-import React, { useState } from "react";
-import {
-  ToolbarItem,
-  Select,
-  SelectVariant,
-  SelectOption,
-} from "@patternfly/react-core";
-import { FilterIcon } from "@patternfly/react-icons";
-import { CompileReportParams } from "@/Core";
-
-interface Props {
-  filterKind: CompileReportParams.Kind | string;
-  setFilterKind: (kind: CompileReportParams.Kind) => void;
-}
-
-export const FilterPicker: React.FC<Props> = ({
-  filterKind,
-  setFilterKind,
-}) => {
-  const [isFilterOpen, setFilterOpen] = useState(false);
-
-  const onSelect = (event, selection) => {
-    setFilterOpen(false);
-    setFilterKind(selection as CompileReportParams.Kind);
-  };
-
-  const items = CompileReportParams.List;
-
-  return (
-    <ToolbarItem>
-      <Select
-        variant={SelectVariant.single}
-        toggleAriaLabel="FilterPicker"
-        onToggle={setFilterOpen}
-        onSelect={onSelect}
-        toggleIcon={<FilterIcon />}
-        selections={filterKind}
-        isOpen={isFilterOpen}
-      >
-        {items.map((item) => (
-          <SelectOption key={item} value={item}>
-            {item}
-          </SelectOption>
-        ))}
-      </Select>
-    </ToolbarItem>
-  );
-};
