--- src/UI/Pages/Parameters/FilterWidget/index.ts
@@ -1,1 +1,0 @@
-export * from "./FilterWidget";
